Updating go-ipld-prime to NodeAssembler era.
The 'testutil/chaintypes' codegen is currently dropped, and I ported
the testchain fixtures to using basicnode instead.  We can restore this
to using codegen soon, but it's easier to complete this set of changes
by temporarily switching to using basicnode.

Tests using 'NewUnencodableSelectorSpec' were discarded -- this more or
less exercised what would happen if an uninitialized zero value got
into a program, and this is now happily statically impossible.

Some tests are not passing; have not sussed yet.