context.TODO() a few places that now require a context
This is due to most of the data pipeline now requiring a context. There is
very few places where this context is needed but wiring a context up to there
quickly grow into a massive change that I'm not able to deal with. I used instead
context.TODO() tofix the compilation while keeping the same behavior as before.