test(responsemanager): remove check
remove checking for cleared requests in test of network errors on block send. When a network error
occurs on a block send, it will error the request and clear it out. However, it is still possible
the execution of the selector query was already complete, in which case, there will not be any
clearing of the request. That's ok.