fix(notifications): fix deadlock on notifications
The notification queue is extremely vulnerable to deadlocks if one of
the subscribers takes a long time to process a notification. This change
makes calls to Subscribe and Publish more non-blocking by treating
internal commands as an unbuffered queue, rather than a blocking
channel.

We also cleanup the subscriber for the response manager. Ultimately, it
may make sense to move these subscribers directly into the message queue
and remove the notification system entirely. but in the meantime, it
makes sense for subscriber not to have access to private vars in the
repsonse manager