Check IPLD context cancellation error type instead of string comparison
Now that upstream issue re error wrapping in IPLD is resolved, use
`errors.Is` to check that a given error does indeed represent context
cancellation. Consolidate duplicate implementation into one exported
utility function.

Note that the functionality of the error check is now different, in that
it no longer match any error that contains the expected string. A test
is added to assert this change as expected behaviour.

Add tests to assert the expected functionality. Note that the tests are
duplicate due to duplicate functionality across the codebase.

See: https://github.com/ipld/go-ipld-prime/issues/58