ipldutil: simplify state synchronization, add docs
The two goroutines were synchronized via two channels,
awaitRequest to signal when StorageReadOpener starts,
and responses to provide it with a value to unblock and return.

Besides those two, there was also stateChan, to pass state around.

Keep the responses channel in place,
but replace awaitRequest and stateChan with a mutex and plain fields.
The added docs describe how they are used and why.

Note that we have to be  careful when locking and unlocking the mutex.
We can't unlock an unlocked mutex, nor leave the mutex locked.
Each call to Lock or Unlock comes with a reasoning.

parentCtx isn't used beyond the constructor, so remove the field.
The visitor and chooser defaults are fallbacks, so use them as such.

Use explicit struct keys for literals, making the code easier to follow.

Fixes #274.