Removes `main` branch callout.
It's been two years since the _checkout to `main`_ instructions were added. I think it's safe to say that folks have switched over their branches by now. It's also commonplace for the primary branch of a repo to be `main` instead of `master`, so calling out this change is likely moot.